requirements.txt
Babel==2.4.0
beautifulsoup4==4.6.0
boto3==1.4.4
botocore==1.5.52
cairocffi==0.8.0
CairoSVG==2.0.2
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
-e .
-e git+git@code.cotidia.com:cotidia/account.git@4ff66e9cb4d252783fe3243b9a79236d20a15276#egg=cotidia_account
-e git+git@code.cotidia.com:cotidia/cms.git@b63a765671f065369ff60c9662134c840962b865#egg=cotidia_cms
-e git+git@code.cotidia.com:cotidia/core.git@8456d0490136fd0509cd3cf9abcda671fcc18b9f#egg=cotidia_core
-e git+git@code.cotidia.com:cotidia/file.git@41c37d5e70ee2d166cca5dcf7b6a557f0fd3792a#egg=cotidia_file
-e git+git@code.cotidia.com:cotidia/mail.git@22f0049a7a423341922817c4e83dc7add4bdd9ab#egg=cotidia_mail
-e git+git@code.cotidia.com:cotidia/meta-data.git@c0217a43a2f82186bc62375e08dceb507bed566a#egg=cotidia_meta_data
-e git+git@code.cotidia.com:cotidia/stripe.git@c84ba4f9d8b64a184dcc7579a8eb7228046b898e#egg=cotidia_stripe
-e git+git@code.cotidia.com:cotidia/team.git@7c89ca23e2b27d5b23b3ae754e0264200d07064a#egg=cotidia_team
cssselect==1.0.1
dj-database-url==0.4.2
Django==2.0.1
django-anymail==1.0
django-appconf==1.0.2
django-betterforms==1.1.4
django-codemirror-widget==0.4.1
django-cors-headers==2.1.0
django-countries==5.3
django-filter==1.0.4
django-form-utils==1.0.3
django-formtools==2.1
django-htmlmin==0.10.0
django-js-asset==0.1.1
django-mptt==0.9.0
django-otp==0.3.13
django-phonenumber-field==1.3.0
django-polymorphic==1.3
django-rest-framework==0.1.0
django-reversion==2.0.10
django-storages==1.5.2
django-two-factor-auth==1.7.0
djangorestframework==3.7.4
djrill==2.1.0
docopt==0.4.0
docutils==0.13.1
factory-boy==2.8.1
Faker==0.7.12
gunicorn==19.7.1
html5lib==0.999999999
idna==2.6
jmespath==0.9.2
jsonfield==1.0.3
lxml==4.2.3
mandrill==1.0.57
olefile==0.44
phonenumberslite==8.4.1
Pillow==4.2.1
pinax-stripe==3.4.1
psycopg2==2.7.1
pycparser==2.17
Pyphen==0.9.4
python-dateutil==2.6.0
python-doc-inherit==0.3.0
python-magic==0.4.13
pytz==2017.3
qrcode==4.0.4
raven==6.4.0
reportlab==3.4.0
requests==2.13.0
s3transfer==0.1.10
six==1.10.0
stripe==1.53.0
text-unidecode==1.0
tinycss==0.4
tqdm==4.19.4
urllib3==1.22
WeasyPrint==0.36
webencodings==0.5.1
whitenoise==3.3.0
